apps/studio/backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
PyYAML==6.0.1


